cleaned some code fixed picture creation (name checked, same name checked as well) pictures (& thumb) are now effectively removed when a Picture is removed from the DB

+++ app/controllers/pictures_controller.rb
@@ -15,11 +15,18 @@ class PicturesController < ApplicationController
   def create
     file_upload = params[:file]
 
-    File.open(path_to_picture(file_upload.original_filename), 'w') do |fh|
-      fh.write(file_upload.read)
-    end
+    if FileTest.exist?(path_to_picture(file_upload.original_filename)) \
+         || FileTest.exist?(path_to_thumbnail(file_upload.original_filename))
 
-    store(file_upload.original_filename, params[:image][:name])
+      flash[:notice] = "This file (or a file with the same name) already exists!"
+      redirect_to :action => 'new'
+    else
+      File.open(path_to_picture(file_upload.original_filename), 'w') do |fh|
+        fh.write(file_upload.read)
+      end
+
+      store(file_upload.original_filename, params[:image][:name])
+    end
   end
 
   def import
@@ -33,9 +40,6 @@ class PicturesController < ApplicationController
   def store(file_to_store, picture_name)
     require 'RMagick'
 
-    # TODO: checks
-    # MIME ?
-
     begin
       thumb = Magick::ImageList.new(path_to_picture(file_to_store))
       thumb.crop_resized!(get_setting('thumb_width').to_i)
@@ -43,7 +47,7 @@ class PicturesController < ApplicationController
     rescue Exception => e
       puts e.message
 
-      flash[:notice] = "This doesn't seem to be a valid picture."
+      flash[:notice] = "This doesn't seem to be a valid picture!"
       redirect_to :action => 'new'
     end
 
@@ -51,10 +55,13 @@ class PicturesController < ApplicationController
       return false;
     end
 
+    if picture_name.empty?
+      picture_name = 'unnamed picture'
+    end
+
     pic = Picture.create(
-      :name => picture_name || 'none',
-      :path => file_to_store,
-      :thumb_path => get_setting('thumb_path_prefix') + file_to_store,
+      :name => picture_name,
+      :filename => file_to_store,
       :user_id => self.current_user.id
     )
     pic.save!
@@ -73,8 +80,12 @@ class PicturesController < ApplicationController
 # - delete associated comments
 #
 
+      FileUtils.rm_f(path_to_thumbnail(@picture.filename))
+      FileUtils.rm_f(path_to_picture(@picture.filename))
+
       @picture.destroy
       @errors = get_errors_for(@picture) if !@picture.errors.empty?
+
       flash[:notice] = ''
       if @errors
         flash[:notice] += @errors
@@ -89,21 +100,31 @@ class PicturesController < ApplicationController
   def show
     @picture = Picture.find(params[:id])
     @user = @picture.user
+
+    render :layout => 'pictures_show'
   end
 
   protected
   def path_to_picture(picture_name)
-    Rails.root.join(
+    path = Rails.root.join(
       get_setting('public_pictures_directory'),
-      picture_name
+      self.current_user.login
     )
+
+    FileUtils.mkdir_p(path) if !File.exist?(path)
+
+    path + picture_name
   end
-  
+
   def path_to_thumbnail(thumbnail_name)
-    Rails.root.join(
+    path = Rails.root.join(
       get_setting('public_pictures_directory'),
       get_setting('thumb_path_prefix'),
-      thumbnail_name
+      self.current_user.login
     )
+
+    FileUtils.mkdir_p(path) if !File.exist?(path)
+
+    path + thumbnail_name
   end
 end
